[ExoBundle] fixes boolean items migration (#5806)
* [ExoBundle] fixes boolean items migration

* Update Updater120403.php
