+++ plugin/exo/Installation/Updater/Updater120403.php
@@ -2,10 +2,7 @@
 
 namespace UJM\ExoBundle\Installation\Updater;
 
-use Claroline\AppBundle\Persistence\ObjectManager;
 use Claroline\InstallationBundle\Updater\Updater;
-use UJM\ExoBundle\Entity\Attempt\Paper;
-use UJM\ExoBundle\Library\Options\Direction;
 
 class Updater120403 extends Updater
 {
@@ -19,7 +16,6 @@ class Updater120403 extends Updater
     public function postUpdate()
     {
         $this->convertDuration();
-        $this->convertBooleanPapers();
         $this->convertBooleanAnswers();
     }
 
@@ -30,66 +26,6 @@ class Updater120403 extends Updater
         $sql = 'UPDATE ujm_exercise SET duration = (duration * 60) WHERE duration IS NOT NULL AND duration != 0';
         $sth = $this->container->get('doctrine.dbal.default_connection')->prepare($sql);
         $sth->execute();
-    }
-
-    private function convertBooleanPapers()
-    {
-        $this->log('Convert boolean papers...');
-
-        /** @var ObjectManager $om */
-        $om = $this->container->get('claroline.persistence.object_manager');
-
-        $papers = $om
-            ->createQuery('
-                SELECT p 
-                FROM UJM\ExoBundle\Entity\Attempt\Paper AS p 
-                WHERE p.structure LIKE :booleanType
-            ')
-            ->setParameters([
-                'booleanType' => '%application/x.boolean+json%',
-            ])
-            ->getResult();
-
-        $i = 0;
-        $total = count($papers);
-
-        /** @var Paper $paper */
-        foreach ($papers as $paper) {
-            ++$i;
-            $this->log("Migrating $i/$total...");
-
-            $structure = json_decode($paper->getStructure(), true);
-            if ($structure && !empty($structure['steps'])) {
-                foreach ($structure['steps'] as $stepIndex => $step) {
-                    if (!empty($step['items'])) {
-                        foreach ($step['items'] as $itemIndex => $item) {
-                            if ('application/x.boolean+json' === $item['type']) {
-                                // replace type
-                                $item['type'] = 'application/x.choice+json';
-
-                                // add missing config
-                                $item['multiple'] = false;
-                                $item['direction'] = Direction::HORIZONTAL;
-
-                                // replace item in structure
-                                $structure['steps'][$stepIndex]['items'][$itemIndex] = $item;
-                            }
-                        }
-                    }
-                }
-
-                $paper->setStructure(json_encode($structure));
-                $om->persist($paper);
-            }
-
-            if (0 === $i % 100) {
-                $om->flush();
-                $this->log('flush');
-            }
-        }
-
-        $om->flush();
-        $this->log('flush');
     }
 
     private function convertBooleanAnswers()
